backend 20-15-57-708/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
python-multipart==0.0.6
slowapi==0.1.9
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
python-dotenv==1.0.0
openai>=1.0.0
anthropic>=0.7.0
google-generativeai>=0.3.0
psutil==5.9.6